+++ source/extensions/isaacsim.core.nodes/premake5.lua
@@ -52,6 +52,7 @@ includedirs {
     "%{root}/source/extensions/isaacsim.core.nodes/include",
     "%{kit_sdk_bin_dir}/dev/fabric/include/",
     "%{root}/_build/target-deps/python/include",
+    "%{root}/source/extensions/isaacsim.robot.schema/include",
 }
 libdirs {
     "%{root}/_build/target-deps/usd/%{cfg.buildcfg}/lib",