+++ source/extensions/isaacsim.robot.schema/premake5.lua
@@ -22,12 +22,11 @@ repo_build.prebuild_link {
     { "data", ext.target_dir .. "/data" },
     { "include", ext.target_dir .. "/include" },
     { "robot_schema", ext.target_dir .. "/usd/schema/isaac/robot_schema" },
+    { "sensor_schema", ext.target_dir .. "/usd/schema/isaac/sensor_schema" },
+    { "range_sensor_schema", ext.target_dir .. "/usd/schema/isaac/range_sensor_schema" },
     { "python/tests", ext.target_dir .. "/isaacsim/robot/schema/tests" },
+    { "python/compat", ext.target_dir .. "/compat" },
 }
 repo_build.prebuild_copy {
     { "python/__init__.py", ext.target_dir .. "/usd/schema/isaac" },
-}
-
-repo_build.prebuild_copy {
-    { "%{root}/_build/target-deps/omni-isaacsim-schema/%{platform}/%{config}", ext.target_dir .. "/plugins" },
 }