+++ source/extensions/isaacsim.sensors.experimental.physics/premake5.lua
@@ -34,9 +34,9 @@ includedirs {
     "plugins/isaacsim.sensors.experimental.physics",
     "%{root}/source/extensions/isaacsim.core.experimental.prims/include",
     "%{root}/source/extensions/isaacsim.core.includes/include",
+    "%{root}/source/extensions/isaacsim.robot.schema/include",
     "%{root}/source/extensions/isaacsim.core.simulation_manager/include",
     target_deps .. "/omni_physics/%{config}/include",
-    "%{root}/_build/target-deps/omni-isaacsim-schema/%{platform}/%{config}/include",
     target_deps .. "/usd/%{cfg.buildcfg}/include",
     target_deps .. "/usd_ext_physics/%{cfg.buildcfg}/include",
     extsbuild_dir .. "/usdrt.scenegraph/include",
@@ -53,11 +53,10 @@ filter {}
 libdirs {
     target_deps .. "/usd/%{cfg.buildcfg}/lib",
     target_deps .. "/usd_ext_physics/%{cfg.buildcfg}/lib",
-    "%{root}/_build/target-deps/omni-isaacsim-schema/%{platform}/%{config}/lib",
     extsbuild_dir .. "/omni.usd.core/bin",
 }
 
-links { "isaacSensorSchema", "omni.usd", "physxSchema" }
+links { "omni.usd", "physxSchema" }
 
 extra_usd_libs = { "usdGeom", "usdPhysics", "usdUtils", "ts" }
 add_usd(extra_usd_libs)